requirements.txt
asgiref==3.6.0
attrs==22.2.0
backports.zoneinfo;python_version<"3.9"
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
djangorestframework-simplejwt==5.2.2
djangorestframework==3.14.0
drf-spectacular==0.26.1
drf-yasg==1.21.5
exceptiongroup==1.1.1
Faker==12.0.1
idna==3.4
importlib-resources==5.12.0
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mixer==7.2.2
myapp==0.1.dev0
packaging==23.0
pkgutil-resolve-name==1.3.10
pluggy==1.0.0
PyJWT==2.6.0
pyrsistent==0.19.3
pytest==7.2.2
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
ruamel.yaml.clib==0.2.7
ruamel.yaml==0.17.21
setuptools==65.6.3
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.15
wheel==0.38.4
zipp==3.15.0
python-dotenv==0.10.1

